+++ src/App.vue
@@ -1,12 +1,10 @@
 <template>
-  <tyh-avatar fit="" src="https://tianyuhao.cn/v3/assets/giraffe.jpg" />
-  <tyh-avatar src="https://tianyuhao.cn2/v3/assets/giraffe.jpg" />
-  <tyh-avatar
-    errorIcon="tyh-ui-githublogo"
-    src="https://tianyuhao.cn2/v3/assets/giraffe.jpg"
-  />
+  <div style="height: 2000px"></div>
+  <tyh-backTop bottom="150px" right="120px" action="smooth">Go</tyh-backTop>
 </template>
 
-<script lang="ts" setup></script>
-
-<style scoped></style>
+<script setup>
+import { ref } from 'vue'
+const radio = ref('right')
+const open = ref(false)
+</script>