+++ src/packages/tyhUi/packages/button/src/type.ts
@@ -1,2 +1,8 @@
-export type buttonType = 'default' | 'primary' | 'success' | 'danger' | 'warning' | ''
+export type buttonType =
+  | 'default'
+  | 'primary'
+  | 'success'
+  | 'danger'
+  | 'warning'
+  | ''
 export type buttonSize = 'large' | 'small' | 'mini' | ''