+++ src/packages/tyhUi/packages/crumbs-item/TyhCrumbsItem.vue
@@ -14,10 +14,10 @@ import { getCurrentInstance } from 'vue'
 const props = defineProps({
   to: String
 })
-const { icon, link } = _TyhCrumbsItem()
-function _TyhCrumbsItem() {
+const { icon, link } = TyhCrumbsItem()
+function TyhCrumbsItem() {
   const icon = ref('')
-  const { proxy } = getCurrentInstance()
+  const { proxy }: any = getCurrentInstance()
   icon.value = inject('Crumbs-separator')
   const link = () => {
     if (!props.to) return