+++ src/packages/tyhUi/packages/division/index.js
@@ -1,4 +1,4 @@
-import T from './TyhDivision.vue'
+import T from './src/TyhDivision.vue'
 
 T.install = _ => _.component('TyhDivision', T)