+++ src/packages/tyhUi/packages/division/src/TyhDivision.vue
@@ -10,23 +10,7 @@
   </div>
 </template>
 
-<script setup>
-defineProps({
-  position: {
-    type: String,
-    default: 'left',
-    validator (v) {
-      return ['left', 'center', 'right'].includes(v)
-    }
-  },
-  color: {
-    type: String,
-    default: '#515a6e'
-  },
-  margin: {
-    type: String,
-    default: '25px'
-  },
-  icon: String
-})
+<script lang="ts" setup>
+import { prop } from './prop'
+defineProps({ ...prop })
 </script>