+++ src/packages/tyhUi/packages/division/src/type.ts
@@ -1,0 +1,1 @@
+export type divisionPosition = 'left' | 'center' | 'right' | ''