+++ src/packages/tyhUi/packages/drawer/src/type.ts
@@ -1,0 +1,1 @@
+export type drawerDirection = 'top' | 'left' | 'bottom' | 'right' | ''