+++ src/packages/tyhUi/packages/icon/TyhIcon.vue
@@ -2,13 +2,13 @@
   <i :class="[`tyh-icon`, icon]" :style="{ color, fontSize: `${size}px` }" />
 </template>
 
-<script setup>
+<script lang="ts" setup>
 defineProps({
   icon: String,
   color: String,
   size: {
-    type: [Number, String],
-    default: '16'
+    type: String,
+    default: () => '16'
   }
 })
 </script>