+++ src/packages/tyhUi/packages/image/index.js
@@ -1,4 +1,4 @@
-import T from './TyhImage.vue'
+import T from './src/TyhImage.vue'
 
 T.install = _ => _.component('TyhImage', T)