+++ src/packages/tyhUi/utils/theme.ts
@@ -1,4 +1,7 @@
-export const THEME: object = {
+interface themeType {
+  [name: string]: string
+}
+export const THEME: themeType = {
   primary: '#3a6ff4',
   success: '#54c600',
   danger: '#d10f1b',